requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
cairocffi @ file:///C:/Users/Administrator/pipwin/cairocffi-1.3.0-cp311-cp311-win_amd64.whl#sha256=3db8f405cc93d3326d867a7482c3f732bd397cb8cfa696c6a4e926c89f7bd9a9
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
docopt==0.6.2
frozenlist==1.4.1
idna==3.6
Js2Py==0.74
multidict==6.0.5
packaging==24.0
pefile==2023.2.7
pillow==10.2.0
pipwin==0.5.2
pycparser==2.21
pygame-ce==2.4.1
pygame-gui==0.6.9
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-i18n==0.3.9
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
tinycss2==1.2.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
yarl==1.9.4
